test+ci(determinism): tighten unit suite to ~4:18 pytest wall
Reduces the determinism unit-test suite from ~5:10 to ~4:18 pytest by
removing coverage that doesn't change the determinism risk surface, plus
small cleanups from the simplify pass. CI-recipe perf path moves the
pretrain args out of the yaml into a checked-in shell script so the
recipe stays small and the experiment is reproducible locally.

Unit suite — tests/unit_tests/determinism/correctness/

* test_hybrid_model.py: scope hybrid parallelism matrix down to one
  PP=2 cell. The first hybrid cell of the suite always pays a ~60s
  Mamba selective_scan + TE attention JIT tax regardless of which
  cell runs first, and that cost dominates the hybrid budget — extra
  parametrize entries add wall time without changing what we verify.
  GPT-model tests still cover TP/EP/FSDP composites in depth.
* test_hybrid_model.py: remove ``_prewarm_mamba_kernels`` autouse
  fixture. The fixture warmed Mamba at TP=1 / pattern="M", but the
  test cells run pattern "M*-" so the JIT for attn+MLP in the hybrid
  layer-spec was never amortized — fixture was paying 23s of setup
  and saving roughly the same amount on the first cell. Net zero.
* test_hybrid_model.py: replace per-method ``BitExactRunner(...).setup()``
  /``.teardown()`` instantiation with a single module-level lifecycle
  helper. The original construct-then-throw-away pattern with a
  ``build_model=lambda *a, **k: None`` placeholder read like a bug.
* test_transformer_layer.py: drop ``@pytest.mark.flaky_in_dev`` from
  the racing-streams and jitter tests. Racing-streams already has
  ``skipif(CUDA_DEVICE_MAX_CONNECTIONS=="1")`` env-var gating; jitter
  passes cleanly in CI. The double-marker had effectively hidden both
  tests from the default ``dev`` CI bucket — a defense-in-depth that
  was hiding working tests.
* test_gpt_model.py: inline the GPT model factory + ``make_gpt_runner``
  helpers that previously lived in ``_gpt_shared.py``. Only
  ``test_fp8_determinism.py`` consumes them; a separate helper module
  for two consumers (one of which is the canonical GPT test file) was
  overkill.
* _gpt_shared.py: deleted.

Perf CI — tests/test_utils/recipes/h100/determinism-perf.yaml

* Slim the recipe to a single ``bash scripts/determinism/perf_breakdown.sh``
  invocation. The pretrain args + nsys profile setup move into the
  checked-in script so the same recipe runs reproducibly locally
  (``bash scripts/determinism/perf_breakdown.sh /tmp/leaderboards /tmp/logs``).

Nsys driver — scripts/determinism/

* run_nsys_breakdown.sh: pass ``--force-export=true`` to ``nsys stats``
  so reruns into a dirty output directory don't fail on stale
  ``.sqlite`` exports older than the new ``.nsys-rep``.
* perf_breakdown.sh: new — holds the pretrain_gpt.py command line that
  used to be inlined in the CI yaml, exports
  ``CUDA_DEVICE_MAX_CONNECTIONS=1`` (Megatron's pre-Blackwell TP>1
  requirement, asserted in arguments.py before --deterministic-mode
  takes effect), and calls run_nsys_breakdown.sh once per mode.
