+++ tests/test_utils/recipes/h100/determinism-perf.yaml
@@ -23,66 +23,10 @@ spec:
     git rev-parse HEAD
   script: |-
     set -euo pipefail
-    # Real pretrain_gpt.py training run, twice: ``nondet`` (BASELINE) and
-    # ``det`` (--deterministic-mode added). The det-vs-nondet loop, nsys
-    # invocation, stats export, and leaderboard join all live in
-    # ``scripts/determinism/run_nsys_breakdown.sh``; this recipe only
-    # provides the pretrain_gpt.py command. Determinism defaults to OFF —
-    # the only difference between the two passes is the
-    # ``--deterministic-mode`` flag injected when
-    # ``DETERMINISM_PERF_MODE=det``.
-    #
-    # CUDA_DEVICE_MAX_CONNECTIONS=1 is Megatron's pre-existing requirement
-    # for TP>1 (asserted at arguments.py:1321 before --deterministic-mode
-    # takes effect). Must be set in the shell.
     cd /opt/megatron-lm
-    export CUDA_DEVICE_MAX_CONNECTIONS=1
-
-    bash scripts/determinism/run_nsys_breakdown.sh \
-      "{assets_dir}/logs/perf-leaderboards" -- \
-      bash -c '
-        set -euo pipefail
-        uv run --no-sync python -m torch.distributed.run \
-          --log-dir "{assets_dir}/logs/torchrun-$DETERMINISM_PERF_MODE" \
-          --tee "0:3,7:3" \
-          --redirects "3" \
-          --nproc_per_node 8 \
-          pretrain_gpt.py \
-            --num-layers 2 \
-            --hidden-size 512 \
-            --num-attention-heads 4 \
-            --seq-length 128 \
-            --max-position-embeddings 128 \
-            --micro-batch-size 2 \
-            --global-batch-size 16 \
-            --train-iters 8 \
-            --lr 1e-4 \
-            --lr-decay-style constant \
-            --lr-decay-iters 100 \
-            --min-lr 1e-5 \
-            --weight-decay 0 \
-            --clip-grad 1.0 \
-            --tensor-model-parallel-size 2 \
-            --pipeline-model-parallel-size 1 \
-            --distributed-backend nccl \
-            --tokenizer-type NullTokenizer \
-            --vocab-size 256 \
-            --mock-data \
-            --split 1,0,0 \
-            --transformer-impl transformer_engine \
-            --use-mcore-models \
-            --no-gradient-accumulation-fusion \
-            --bf16 \
-            --log-interval 1 \
-            --eval-iters 0 \
-            --eval-interval 10000 \
-            --no-load-optim \
-            --no-load-rng \
-            $([ "$DETERMINISM_PERF_MODE" = det ] && echo --deterministic-mode) \
-            --profile \
-            --profile-step-start 4 \
-            --profile-step-end 7
-      '
+    bash scripts/determinism/perf_breakdown.sh \
+      "{assets_dir}/logs/perf-leaderboards" \
+      "{assets_dir}/logs"
 
 products:
   # Routed through the integration-tests matrix (scope: mr-github) — picked
