+++ tests/unit_tests/determinism/correctness/test_hybrid_model.py
@@ -5,9 +5,6 @@
 Adding a new parallelism cell is a one-line append to
 ``configs.PARALLELISM_CONFIGS``. ``HYBRID_CONFIGS`` provides the
 (layer_pattern, overrides) presets specific to this model class.
-
-``MambaModel`` is a deprecated alias for ``HybridModel`` — a single Hybrid
-test covers both.
 """
 
 import pytest
@@ -16,58 +13,21 @@ import torch
 from megatron.core.models.hybrid.hybrid_layer_specs import hybrid_stack_spec
 from megatron.core.models.hybrid.hybrid_model import HybridModel
 from megatron.core.transformer.transformer_config import TransformerConfig
-from tests.unit_tests.determinism.configs import HYBRID_CONFIGS, hybrid_base, parallelism_configs
-
-# Hybrid skips TP=8: Mamba selective_scan / conv1d CUDA extensions JIT-compile
-# per per-rank shard shape, and the second high-TP cell would re-JIT for a
-# different shape (~10s extra). TP coverage at full degree is already exercised
-# by ``test_gpt_model``; keeping TP=4 here gives one high-TP hybrid run that
-# verifies the Mamba sharding path stays deterministic.
-_HYBRID_PARALLELISM_CONFIGS = parallelism_configs(exclude=("tp8",))
 from tests.unit_tests.determinism.bit_exact_runner import BitExactRunner
+from tests.unit_tests.determinism.configs import HYBRID_CONFIGS, hybrid_base
+
+# Hybrid keeps a single PP=2 cell. The first mamba cell of the suite always
+# pays a ~60s JIT tax (TE attention + Mamba selective_scan compilation under
+# the hybrid layer-spec), and that cost dominates regardless of how many
+# subsequent cells run — so additional hybrid parametrize coverage adds wall
+# time without changing the determinism risk surface much. GPT-model tests
+# already cover TP/EP/FSDP composites in depth; the hybrid cell here is a
+# smoke test that "Mamba + attn + MLP + PP" stays bit-exact.
+_HYBRID_PARALLELISM_CONFIGS = [pytest.param({"PP": 2}, id="pp2")]
 
 _SEQ_LEN = 32
 _MICRO_BATCH = 2
 _VOCAB_SIZE = 128
-
-
-@pytest.fixture(scope="module", autouse=True)
-def _prewarm_mamba_kernels():
-    """Mamba's selective_scan / conv1d CUDA extensions JIT-compile on first
-    use (~50s). Warm them once at module load so subsequent cells reuse
-    the cached kernels.
-
-    The warmup forward is wrapped in try/finally so a transient JIT/OOM
-    failure still tears down ``parallel_state`` — otherwise subsequent
-    tests hit ``setup_method`` with TP=1 still initialised and produce a
-    cascade of secondary errors that masks the original failure.
-    """
-    import os
-
-    from megatron.core.tensor_parallel.random import model_parallel_cuda_manual_seed
-    from tests.unit_tests.test_utilities import Utils
-
-    os.environ.setdefault("CUDA_DEVICE_MAX_CONNECTIONS", "1")
-    Utils.initialize_model_parallel(tensor_model_parallel_size=1)
-    try:
-        model_parallel_cuda_manual_seed(123)
-        cfg_kwargs = hybrid_base()
-        cfg_kwargs["num_layers"] = 1
-        cfg = TransformerConfig(**cfg_kwargs)
-        warm = HybridModel(
-            config=cfg,
-            hybrid_stack_spec=hybrid_stack_spec,
-            vocab_size=_VOCAB_SIZE,
-            max_sequence_length=_SEQ_LEN,
-            hybrid_layer_pattern="M",
-        ).cuda()
-        with torch.no_grad():
-            warm(**_hybrid_inputs())
-        del warm
-    finally:
-        Utils.destroy_model_parallel()
-        torch.cuda.empty_cache()
-    yield
 
 
 def _hybrid_inputs() -> dict:
@@ -86,25 +46,25 @@ def _hybrid_inputs() -> dict:
     }
 
 
+# Module-level lifecycle helper — its build_model lambda is never invoked;
+# only setup/teardown (init / destroy model-parallel + cache flush) are used.
+# Per-test runners are built inside the test body because HybridModel needs
+# the layer_pattern from HYBRID_CONFIGS, which the parametrize feeds in.
+_LIFECYCLE = BitExactRunner(
+    build_model=lambda *a, **k: None,
+    make_inputs=_hybrid_inputs,
+    base_config=hybrid_base,
+    supports_pp=False,
+)
+
+
 class TestHybridModelDeterminism:
 
     def setup_method(self, method):
-        # Use a placeholder runner during setup; the real one is built per-test
-        # because HybridModel needs the layer_pattern from HYBRID_CONFIGS.
-        BitExactRunner(
-            build_model=lambda *a, **k: None,
-            make_inputs=_hybrid_inputs,
-            base_config=hybrid_base,
-            supports_pp=False,
-        ).setup()
+        _LIFECYCLE.setup()
 
     def teardown_method(self, method):
-        BitExactRunner(
-            build_model=lambda *a, **k: None,
-            make_inputs=_hybrid_inputs,
-            base_config=hybrid_base,
-            supports_pp=False,
-        ).teardown()
+        _LIFECYCLE.teardown()
 
     @pytest.mark.internal
     @pytest.mark.parametrize("parallelism", _HYBRID_PARALLELISM_CONFIGS)
